% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaCInG-package.R
\docType{package}
\name{RaCInG-package}
\alias{RaCInG}
\alias{RaCInG-package}
\title{RaCInG: Random Cell-cell Interaction Generator}
\description{
\code{RaCInG} provides functions to prepare ligand-receptor input matrices,
generate patient-specific communication graphs, derive kernel or Monte Carlo
features, and perform downstream statistical analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VeraPancaldiLab/RaCInG_package}
  \item \url{https://VeraPancaldiLab.github.io/RaCInG_package/}
  \item Report bugs at \url{https://github.com/VeraPancaldiLab/RaCInG_package/issues}
}

}
\author{
\strong{Maintainer}: Marcelo Hurtado \email{marcelo.hurt137@gmail.com}

}
\keyword{internal}
